pyproject.toml
[build-system]
requires = [
    "setuptools >= 61.2",
    "setuptools_scm[toml] >= 6.2"
]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
# Presence of this section activates setuptools_scm, even though it's empty

[project]
name = "ebrick_demo"
description = "ebrick-demo"
version = "0.0.1"
dependencies = [
    "switchboard-hw == 0.0.38",
    "siliconcompiler == 0.21.4",
    "umi @ git+https://github.com/zeroasiccorp/umi@main",
    "lambdalib == 0.2.0"
]

[project.optional-dependencies]
# Test dependencies.
test = [
    "flake8 == 7.0.0",
    "tclint == 0.2.2"
]

[tool.setuptools]
include-package-data = true
packages = [
    "ebrick_demo"
]

[tool.tclint]
exclude = [
    'build/',
    'ebrick/build/'
]

[tool.tclint.style]
allow-aligned-sets = true
line-length = 80
indent = 4
max-blank-lines = 1
